diff_simple.py
Easily find the def differences between branches of the same file.

Place this script in the root of the repo, make it executable, then run it with <path/to/file> <old/branch> <new/branch>.

It outputs two CSV files. One contains a list of all new or updated definitions, and the other a list of all removed definitions. Each new or updated def will be noted with git blame.